requirements.txt
qrcode
numpy
Pillow
transformers
torch
torchvision
pandas
